YoutubeStreamNotifyBot/requirements.txt
loguru
tweepy
trio
python-telegram-bot
google-api-python-client
python-dateutil
requests
